- Smoke Test Seg Fault fix
(cherry picked from commit 550c055010e27dc0f02a421b378330a763d6b479)

# Conflicts:
#	geonode/layers/tests.py
